refactor: remove genkit ai flows
