--- src/ai/genkit.ts
@@ -1,7 +1,0 @@
-import {genkit} from 'genkit';
-import {googleAI} from '@genkit-ai/googleai';
-
-export const ai = genkit({
-  plugins: [googleAI({useGoogleAuth: true})],
-  model: 'googleai/gemini-2.0-flash',
-});
