+++ src/app/clientes/components/client-list.tsx
@@ -40,8 +40,8 @@ import { Label } from '@/components/ui/label';
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
 import { Separator } from '@/components/ui/separator';
-import { settleCustomerPayment } from '@/ai/flows/settle-customer-payment';
-import { registerCustomer } from '@/ai/flows/register-customer';
+import { settleCustomerPayment } from '@/services/settle-customer-payment';
+import { registerCustomer } from '@/services/register-customer';
 
 
 type FilterTab = 'all' | 'pessoa-juridica' | 'pessoa-fisica' | 'com-pendencias';
