+++ src/services/register-manual-purchase.ts
@@ -1,0 +1,107 @@
+// @ts-nocheck
+'use server';
+
+import { z } from 'zod';
+import { db, collection, addDoc, doc, runTransaction, getDoc, increment } from '@/lib/netly';
+import type { Purchase, FinancialMovement, Supplier, Ingredient, SourceAccount } from '@/lib/types';
+import { format, addMonths } from 'date-fns';
+
+const PurchaseItemSchema = z.object({
+  ingredientId: z.string(),
+  quantity: z.number(),
+  unitPrice: z.number(),
+});
+
+const RegisterManualPurchaseInputSchema = z.object({
+  supplierId: z.string(),
+  invoiceNumber: z.string().optional(),
+  date: z.string(),
+  sourceAccount: z.custom<SourceAccount>(),
+  paymentMethod: z.enum(['pix', 'boleto', 'dinheiro', 'cartao_credito', 'cartao_debito']),
+  installments: z.number().int().min(1),
+  firstDueDate: z.string(),
+  items: z.array(PurchaseItemSchema),
+  totalAmount: z.number(),
+});
+export type RegisterManualPurchaseInput = z.infer<typeof RegisterManualPurchaseInputSchema>;
+
+export async function registerManualPurchase(
+  input: RegisterManualPurchaseInput
+): Promise<{ purchaseId: string; message: string }> {
+  const supplierDoc = await getDoc(doc(db, 'suppliers', input.supplierId));
+  if (!supplierDoc.exists()) {
+    throw new Error('Fornecedor não encontrado.');
+  }
+  const supplier = supplierDoc.data() as Supplier;
+
+  const purchaseId = await runTransaction(db, async (transaction) => {
+    const purchaseData: Omit<Purchase, 'id'> = {
+      supplierId: input.supplierId,
+      invoiceNumber: input.invoiceNumber || `MANUAL-${Date.now()}`,
+      date: input.date,
+      totalAmount: input.totalAmount,
+      paymentMethod: input.paymentMethod,
+      items: [],
+    };
+
+    const purchaseRef = doc(collection(db, 'purchases'));
+
+    for (const item of input.items) {
+      const ingredientRef = doc(db, 'ingredients', item.ingredientId);
+      const ingredientDoc = await transaction.get(ingredientRef);
+      if (!ingredientDoc.exists()) {
+        throw new Error(`Insumo com ID ${item.ingredientId} não encontrado.`);
+      }
+
+      const ingredientData = ingredientDoc.data() as Ingredient;
+      const oldStock = ingredientData.stock;
+      const oldCost = ingredientData.cost;
+      const newQuantity = item.quantity;
+      const newPrice = item.unitPrice;
+      const newTotalStock = oldStock + newQuantity;
+      const newAverageCost = newTotalStock > 0
+        ? ((oldStock * oldCost) + (newQuantity * newPrice)) / newTotalStock
+        : newPrice;
+
+      transaction.update(ingredientRef, {
+        stock: increment(newQuantity),
+        cost: newAverageCost,
+      });
+
+      purchaseData.items.push({
+        name: ingredientData.name,
+        quantity: item.quantity,
+        unitPrice: item.unitPrice,
+      });
+    }
+
+    transaction.set(purchaseRef, purchaseData);
+
+    const installmentValue = input.totalAmount / input.installments;
+    const isPaidOnPurchase = ['pix', 'dinheiro', 'cartao_debito'].includes(input.paymentMethod);
+
+    for (let i = 0; i < input.installments; i++) {
+      const dueDate = addMonths(new Date(input.firstDueDate), i);
+      const financialMovement: Omit<FinancialMovement, 'id'> = {
+        description: `Compra ${purchaseData.invoiceNumber} - ${supplier.name} (Parc. ${i + 1}/${input.installments})`,
+        referenceId: purchaseRef.id,
+        dueDate: format(dueDate, 'yyyy-MM-dd'),
+        amount: -installmentValue,
+        status: isPaidOnPurchase ? 'paid' : 'pending',
+        paymentDate: isPaidOnPurchase ? format(new Date(input.date), 'yyyy-MM-dd') : undefined,
+        category: 'insumos',
+        sourceAccount: input.sourceAccount,
+        type: 'expense',
+      };
+      const movementRef = doc(collection(db, 'financialMovements'));
+      transaction.set(movementRef, financialMovement);
+    }
+
+    return purchaseRef.id;
+  });
+
+  return {
+    purchaseId,
+    message: `Compra de ${supplier.name} no valor de ${input.totalAmount.toLocaleString('pt-BR', { style: 'currency', currency: 'BRL' })} registrada com sucesso!`,
+  };
+}
