+++ src/services/register-sale.ts
@@ -1,0 +1,117 @@
+// @ts-nocheck
+'use server';
+
+import { z } from 'zod';
+import { db, collection, doc, runTransaction, increment } from '@/lib/netly';
+import type { Product, FinancialMovement, Customer, SourceAccount, Sale, SaleItem, SaleChannel } from '@/lib/types';
+import { format } from 'date-fns';
+
+const SaleItemSchema = z.object({
+  productId: z.string(),
+  quantity: z.number(),
+  unitPrice: z.number(),
+  productName: z.string(),
+});
+
+const RegisterSaleInputSchema = z.object({
+  items: z.array(SaleItemSchema),
+  totalAmount: z.number(),
+  paymentMethod: z.enum(['pix', 'boleto', 'dinheiro', 'cartao_credito', 'cartao_debito']),
+  sourceAccount: z.custom<SourceAccount>(),
+  customerId: z.string().optional(),
+  dueDate: z.string().optional(),
+  status: z.enum(['concluida', 'pendente']),
+  channel: z.custom<SaleChannel>(),
+  salespersonId: z.string().optional(),
+  location: z.string().optional(),
+  notes: z.string().optional(),
+});
+export type RegisterSaleInput = z.infer<typeof RegisterSaleInputSchema>;
+
+export async function registerSale(
+  input: RegisterSaleInput
+): Promise<{ message: string; saleId: string }> {
+  const { items, totalAmount, paymentMethod, sourceAccount, customerId, dueDate, channel, salespersonId, location, notes, status } = input;
+  const isSaleOnCredit = !!dueDate;
+
+  const { saleId, productNames } = await runTransaction(db, async (transaction) => {
+    const productNames: string[] = [];
+    const isConcluded = status === 'concluida';
+
+    if (isConcluded) {
+      for (const item of items) {
+        const productRef = doc(db, 'products', item.productId);
+        const productDoc = await transaction.get(productRef);
+        if (!productDoc.exists()) {
+          throw new Error(`Produto ${item.productName} não encontrado.`);
+        }
+        const product = productDoc.data() as Product;
+        const currentStock = product.stock || 0;
+        if (currentStock < item.quantity) {
+          throw new Error(`Estoque insuficiente para ${product.name}. Disponível: ${currentStock}, Solicitado: ${item.quantity}`);
+        }
+        transaction.update(productRef, {
+          stock: increment(-item.quantity),
+          sold: increment(item.quantity),
+        });
+        productNames.push(item.productName);
+      }
+    }
+
+    const today = new Date();
+    const saleRef = doc(collection(db, 'sales'));
+    const saleData: Omit<Sale, 'id'> = {
+      customerId,
+      date: format(today, 'yyyy-MM-dd'),
+      items: items.map(i => ({ productId: i.productId, productName: i.productName, quantity: i.quantity, unitPrice: i.unitPrice })),
+      totalAmount,
+      paymentMethod,
+      status,
+      channel,
+      salespersonId: salespersonId === 'none' ? undefined : salespersonId,
+      location,
+      notes,
+    };
+    transaction.set(saleRef, saleData);
+
+    if (isConcluded) {
+      const movementStatus = isSaleOnCredit ? 'pending' : 'paid';
+      const movementDescription = `Venda ${saleRef.id}: ${items.length} item(s) - ${productNames.slice(0, 2).join(', ')}${productNames.length > 2 ? '...' : ''}`;
+      const financialMovement: Omit<FinancialMovement, 'id'> = {
+        description: movementDescription,
+        referenceId: saleRef.id,
+        customerId,
+        dueDate: isSaleOnCredit ? dueDate : format(today, 'yyyy-MM-dd'),
+        paymentDate: movementStatus === 'paid' ? format(today, 'yyyy-MM-dd') : undefined,
+        amount: totalAmount,
+        status: movementStatus,
+        type: 'revenue',
+        category: 'vendas',
+        sourceAccount,
+      };
+      const movementRef = doc(collection(db, 'financialMovements'));
+      transaction.set(movementRef, financialMovement);
+    }
+
+    if (customerId && isConcluded) {
+      const customerRef = doc(db, 'customers', customerId);
+      const customerDoc = await transaction.get(customerRef);
+      if (!customerDoc.exists()) {
+        throw new Error(`Cliente ${customerId} não encontrado.`);
+      }
+      transaction.update(customerRef, {
+        pendingAmount: isSaleOnCredit ? increment(totalAmount) : increment(0),
+        lastPurchaseDate: format(today, 'dd/MM/yyyy'),
+        totalOrders: increment(1),
+        totalPurchasesValue: increment(totalAmount),
+      });
+    }
+
+    return { saleId: saleRef.id, productNames };
+  });
+
+  return {
+    message: `Venda ${saleId} no valor de ${totalAmount.toLocaleString('pt-BR', { style: 'currency', currency: 'BRL' })} registrada com sucesso.`,
+    saleId,
+  };
+}
